+++ html/src/Application/Process/ProcessInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Otus\Food\Application\Process;
+
+use Otus\Food\Domain\Kitchen\Entity\Meal;
+
+interface ProcessInterface
+{
+    /**
+     * @return Meal
+     */
+    public function run(): Meal;
+}
