+++ html/src/Application/Strategy/Burger.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Otus\Food\Application\Strategy;
 
+use Otus\Food\Application\Process\ProcessBurger;
 use Otus\Food\Application\UseCase\CookingBurger;
 use Otus\Food\Domain\Kitchen\Entity\Meal;
 use Otus\Food\Domain\Order\Order;
@@ -17,6 +18,6 @@ final class Burger implements MealStrategy
      */
     public function cooking(Order $order): Meal
     {
-        return new CookingBurger($order)->cooking();
+        return new CookingBurger(new ProcessBurger($order))->cooking();
     }
 }
