+++ html/src/Application/Strategy/Pizza/Chef.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Otus\Food\Application\Strategy\Pizza;
 
+use Otus\Food\Application\Process\Pizza\ChefProcess;
 use Otus\Food\Application\Strategy\MealStrategy;
 use Otus\Food\Application\UseCase\Pizza\CookingChef;
 use Otus\Food\Domain\Kitchen\Entity\Meal;
@@ -18,6 +19,6 @@ final class Chef implements MealStrategy
      */
     public function cooking(Order $order): Meal
     {
-        return new CookingChef($order)->cooking();
+        return new CookingChef(new ChefProcess($order))->cooking();
     }
 }
