+++ html/src/Application/Strategy/Sandwich.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Otus\Food\Application\Strategy;
 
+use Otus\Food\Application\Process\ProcessSandwich;
 use Otus\Food\Application\UseCase\CookingSandwich;
 use Otus\Food\Domain\Kitchen\Entity\Meal;
 use Otus\Food\Domain\Order\Order;
@@ -17,6 +18,6 @@ final class Sandwich implements MealStrategy
      */
     public function cooking(Order $order): Meal
     {
-        return new CookingSandwich($order)->cooking();
+        return new CookingSandwich(new ProcessSandwich($order))->cooking();
     }
 }
