+++ html/src/Application/UseCase/CookingSandwich.php
@@ -4,20 +4,15 @@ declare(strict_types=1);
 
 namespace Otus\Food\Application\UseCase;
 
-use Otus\Food\Application\Builder\Proxy;
-use Otus\Food\Application\Builder\Sandwich;
-use Otus\Food\Application\Factory\SandwichFactory;
-use Otus\Food\Application\Observer\Kitchen;
-use Otus\Food\Application\Observer\Monitor;
+use Otus\Food\Application\Process\ProcessInterface;
 use Otus\Food\Domain\Kitchen\Entity\Meal;
-use Otus\Food\Domain\Order\Order;
 
-final class CookingSandwich
+final readonly class CookingSandwich
 {
     /**
-     * @param Order $order
+     * @param ProcessInterface $process
      */
-    public function __construct(private Order $order)
+    public function __construct(private ProcessInterface $process)
     {
     }
 
@@ -26,16 +21,6 @@ final class CookingSandwich
      */
     public function cooking(): Meal
     {
-        $meal = SandwichFactory::factory();
-
-        $builder = new Sandwich($meal, $this->order);
-
-        // todo
-        $kitchen = new Kitchen();
-        $kitchen->attach(new Monitor());
-
-        $proxy = new Proxy($builder, $kitchen);
-
-        return $proxy->cooking();
+        return $this->process->run();
     }
 }
